Remove BookSeeder from DatabaseSeeder
This change updates the DatabaseSeeder by removing the BookSeeder class from the list of seeders. This might be done to prevent duplication or because BookSeeder is now obsolete or handled differently. Ensure to run the seeder and test to verify that the change does not affect the data population process.